feat: implement fare estimation and ETA preview
- Fix type mismatch: RideFareCalculationStrategy returns double, convert to BigDecimal
- Add FareEstimateRequestDto and FareEstimateDto
- Add FareEstimationService interface and implementation
- Add calculateDistanceAndDuration to DistanceService (parse duration from OSRM)
- Add POST /riders/fare-estimate endpoint to RiderController

Test: 93 tests passed

+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/services/impl/ScheduledRideServiceImpl.java
@@ -52,8 +52,8 @@ public class ScheduledRideServiceImpl implements ScheduledRideService {
         rideRequest.setRideRequestStatus(RideRequestStatus.SCHEDULED);
         rideRequest.setScheduledTime(scheduledTime);
 
-        BigDecimal fare = rideStrategyManager.rideFareCalculationStrategy().calculateFare(rideRequest);
-        rideRequest.setFare(fare);
+        Double fare = rideStrategyManager.rideFareCalculationStrategy().calculateFare(rideRequest);
+        rideRequest.setFare(BigDecimal.valueOf(fare));
 
         RideRequest saved = rideRequestRepository.save(rideRequest);
         log.info("Scheduled ride created: id={}, scheduledTime={}", saved.getId(), scheduledTime);
@@ -81,8 +81,8 @@ public class ScheduledRideServiceImpl implements ScheduledRideService {
         RideRequest rideRequest = findOwnedScheduledRequest(rideRequestId);
         rideRequest.setScheduledTime(newScheduledTime);
 
-        BigDecimal fare = rideStrategyManager.rideFareCalculationStrategy().calculateFare(rideRequest);
-        rideRequest.setFare(fare);
+        Double fare = rideStrategyManager.rideFareCalculationStrategy().calculateFare(rideRequest);
+        rideRequest.setFare(BigDecimal.valueOf(fare));
 
         RideRequest saved = rideRequestRepository.save(rideRequest);
         log.info("Scheduled ride rescheduled: id={}, newTime={}", rideRequestId, newScheduledTime);
